# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My DOC api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Show all Blogs",
                "operationId": "79fd4c40fed6c98024ab5da0e69c56d0",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference for the response. Default is 'en'. Possible values: 'en', 'ru', 'kg', 'kz', 'tj', 'tm', 'uz'.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ru",
                                "kg",
                                "kz",
                                "tj",
                                "tm",
                                "uz"
                            ],
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of blogs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "title 1 for 1"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "test.png"
                                                    },
                                                    "destination": {
                                                        "type": "string",
                                                        "example": "Кыргызстан"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "test description 1 for 1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No blogs found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "not_found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No blogs found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/blogs/{blog}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Show Blog",
                "operationId": "2881790c4df1629b0f19cb78a59b6a76",
                "parameters": [
                    {
                        "name": "blog",
                        "in": "path",
                        "description": "ID of the blog to retrieve",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language preference for the response. Default is 'en'. Possible values: 'en', 'ru', 'kg', 'kz', 'tj', 'tm', 'uz'.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ru",
                                "kg",
                                "kz",
                                "tj",
                                "tm",
                                "uz"
                            ],
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of the blog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "title 1 for 1"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "test.png"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "Кыргызстан"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "test description 1 for 1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Blog not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "not_found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Blog not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}